refactor(mobile): replace TabScreen with common wrapper
