+++ apps/mobile/src/lib/navigation/bottom-tab/native.android.tsx
@@ -2,3 +2,5 @@ import { requireNativeView } from "expo"
 import type { ViewProps } from "react-native"
 
 export const TabBarPortalWrapper = requireNativeView<ViewProps>("TabBarPortal")
+
+export const TabScreenWrapper = requireNativeView<ViewProps>("TabScreen")
