+++ apps/mobile/src/lib/navigation/bottom-tab/native.tsx
@@ -1,1 +1,2 @@
 export { View as TabBarPortalWrapper } from "react-native"
+export { View as TabScreenWrapper } from "react-native"
